+++ test/unit/adobe_air_test.rb
@@ -9,7 +9,7 @@ class AdobeAirTest < Minitest::Test
     assert browser.webkit?
     assert_equal "0", browser.version
     assert_equal "0.0", browser.full_version
-    assert_equal "Generic Browser", browser.name
+    assert_equal "Unknown Browser", browser.name
     assert_equal "Adobe AIR", browser.platform.name
     assert_equal "13.0", browser.platform.version
   end