+++ lib/browser/browser.rb
@@ -33,6 +33,7 @@ require_relative "snapchat"
 require_relative "duck_duck_go"
 require_relative "samsung_browser"
 require_relative "huawei_browser"
+require_relative "miui_browser"
 require_relative "maxthon"
 require_relative "google_search_app"
 
@@ -79,6 +80,7 @@ module Browser
       DuckDuckGo,           # must be placed before Chrome and Safari
       SamsungBrowser,       # must be placed before Chrome and Safari
       HuaweiBrowser,        # must be placed before Chrome and Safari
+      MiuiBrowser,          # must be placed before Chrome and Safari
       Maxthon,              # must be placed before Chrome and Safari
       GoogleSearchApp,      # must be placed before Chrome and Safari
       Chrome,