+++ CHANGELOG.md
@@ -9,6 +9,14 @@
 - All the changes above affect how `browser.meta` is composed
 - Add method `Browser::Base#unknown?`
 - Fix issue with `Browser::Base#safari?` matching full version
+- Add FullStory to the bot list.
+- Add Charlotte to the bot list.
+- Add Echobox to the bot list.
+- Add Yandex's search browser to the exception list.
+- Add Cubot devices to the exception list.
+- Add NAVER search app to the exception list.
+- Add MiuiQuickSearchBox to the exception list.
+- Add MEIZU_AppSearch to the exception list.
 
 ## 4.2.0