+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Add Maxthon detection
 - Add Google Search App detection
 - Add Huawei Browser detection
+- Fix Duck Duck Go browser that was being recognized as a bot.
 
 ## 4.2.0