+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Add Google Search App detection
 - Add Huawei Browser detection
 - Fix Duck Duck Go browser that was being recognized as a bot.
+- Add Miui Browser detection
 
 ## 4.2.0